Add todos base api functionality

+++ src/controller/project-controller.ts
@@ -5,7 +5,7 @@ import {Project} from "../model/project";
 export const createProject: RequestHandler = async (req, res, next) => {
     let project = req.body as Project;
     const newProject = await projectService.saveProject(project);
-    return res.json(newProject).status(200);
+    return res.status(200).json(newProject);
 }
 
 export const getProject: RequestHandler = async (req, res, next) => {
@@ -15,7 +15,7 @@ export const getProject: RequestHandler = async (req, res, next) => {
     const project = await projectService.findProjectById(projectId);
     if (project) {
         console.log(project);
-        return res.json(project).status(200);
+        return res.status(200).json(project);
     }
 
     return res.status(404).json({message: `Project with id ${projectId} not found.`});
@@ -23,13 +23,19 @@ export const getProject: RequestHandler = async (req, res, next) => {
 
 export const getAllProject: RequestHandler = async (req, res, next) => {
     const projects: Project[] = await projectService.findAllProjects();
-    return res.json(projects).status(200);
+    return res.status(200).json(projects);
 }
 
 export const deleteProject: RequestHandler = async (req, res, next) => {
     const projectId = +req.params.id;
-    await projectService.deleteProject(projectId)
-    return res.json({message: `Project with id ${projectId} removed.`}).status(200);
+    const project = await projectService.findProjectById(projectId);
+
+    if (project) {
+        await projectService.deleteProject(projectId)
+        return res.status(200).json({message: `Project with id ${projectId} removed.`});
+    }
+
+    return res.status(404).json({message: `Project with id ${projectId} not found.`});
 }
 
 export const updateProject: RequestHandler = async (req, res, next) => {
@@ -40,7 +46,7 @@ export const updateProject: RequestHandler = async (req, res, next) => {
 
     if (project) {
         await projectService.updateProject(projectId, updateProject);
-        return res.json({message: 'updated'}).status(200);
+        return res.status(200).json({message: 'updated'});
     }
 
     return res.status(404).json({message: `Project with id ${projectId} not found.`});
